Replaced deprecated mysql commands. CMS should now 'work'.

+++ _install.php
@@ -1,9 +1,16 @@
 <?php
-$linkID = mysql_connect("localhost", "jumento", "pass");
-mysql_select_db("jumento", $linkID);
+$linkID = new mysqli("localhost", "root", "password","jumento");
+if($linkID->connect_errno){
+	printf("Connection failed %s\n",$linkID.connect_error);
+	exit();
+}
+//mysql_select_db("timparki_jumento", $linkID);
 $droppages= 'DROP TABLE IF EXISTS `pages`';
 echo $droppages."<br>";
-$droppagesresult = mysql_query ($droppages, $linkID) or die("Drop Content Failed: ".mysql_error());
+if($linkID->query ($droppages) ===TRUE){
+  printf("Pages dropped successfully.\n");
+}
+// or die("Drop Content Failed: ".mysql_error());
 $createpages = "CREATE TABLE IF NOT EXISTS `pages` (
   `page_id` mediumint(9) NOT NULL auto_increment,
   `parent_id` mediumint(9) NOT NULL default '0',
@@ -12,17 +19,24 @@ $createpages = "CREATE TABLE IF NOT EXISTS `pages` (
   `page_description` varchar(255) NOT NULL default 'default description',
   `page_keywords` varchar(255) NOT NULL default 'default keywords',
   `user_created` varchar(100) NOT NULL default 'admin',
-  `date_created` timestamp NOT NULL default '0000-00-00 00:00:00',
+  `date_created` timestamp NOT NULL default CURRENT_TIMESTAMP,
   `user_modified` varchar(100) NOT NULL default 'admin',
   `date_modified` timestamp NOT NULL default CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP,
   `page_visibility` enum('public','private','group') NOT NULL default 'public',
   PRIMARY KEY  (`page_id`)
 )";
 echo $createpages ."<br>";
-$createpagesresult = mysql_query($createpages, $linkID) or die("Create Content Failed: ". mysql_error());
+
+if($linkID->query($createpages) ===TRUE){
+  printf("Pages created successfully.\n");
+}else{
+   printf($linkID->error ,"\n\n");
+}
+
+// or die("Create Content Failed: ". mysql_error());
 $dropsections = "DROP TABLE IF EXISTS `sections`";
 echo $dropsections."<br>";
-$dropsectionsresult = mysql_query($dropsections, $linkID) or die("Drop Sections Failed: ".mysql_error());
+$dropsectionsresult = $linkID->query ($dropsections);
 $createsections="CREATE TABLE IF NOT EXISTS `sections` (
   `section_id` mediumint(9) NOT NULL auto_increment,
   `page_id` mediumint(9) NOT NULL default '0',
@@ -30,17 +44,24 @@ $createsections="CREATE TABLE IF NOT EXISTS `sections` (
   `section_type` enum('html','wysiwyg_basic','wysiwyg_full') default 'html',
   `section_content` text,
   `section_user_created` varchar(100) NOT NULL default '0',
-  `section_date_created` timestamp NOT NULL default '0000-00-00 00:00:00',
+  `section_date_created` timestamp NOT NULL default CURRENT_TIMESTAMP,
   `section_user_modified` varchar(100) NOT NULL default '0',
   `section_date_modified` timestamp NOT NULL default CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP,
   `section_status` enum('active','inactive') NOT NULL default 'active',
   PRIMARY KEY  (`section_id`)
 )";
 echo $createsections."<br>";
-$createsectionsresult = mysql_query($createsections, $linkID) or die("Create sections failed: ".mysql_error());
+
+if($linkID->query($createsections) ===TRUE){
+  printf("Sections created successfully.\n");
+}else{
+   printf($linkID->error ,"\n\n");
+}
+
+
 $dropusers = "DROP TABLE IF EXISTS `users`";
 echo $dropusers."<br>";
-$dropusersresult = mysql_query($dropusers, $linkID) or die("Drop users Failed: ".mysql_error());
+$dropusersresult = $linkID->query ($dropusers);
 $createusers="CREATE TABLE IF NOT EXISTS `users` (
   `user_id` mediumint(9) NOT NULL auto_increment,
   `user_name` varchar(100) NOT NULL default '',
@@ -49,13 +70,19 @@ $createusers="CREATE TABLE IF NOT EXISTS `users` (
   `user_group` enum('1','2','3') NOT NULL default '1',
   PRIMARY KEY  (`user_id`),
   UNIQUE KEY `user_name` (`user_name`)
-)TYPE=MyISAM AUTO_INCREMENT=5";
+)AUTO_INCREMENT=5";
 echo $createusers."<br>";
-$createuserssresult = mysql_query($createusers, $linkID) or die("Create users failed: ".mysql_error());
+
+if($linkID->query($createusers) ===TRUE){
+  printf("Users created successfully.\n");
+}else{
+   printf($linkID->error ,"\n\n");
+}
+
 
 $dropgroups = "DROP TABLE IF EXISTS `group_permissions`";
 echo $dropgroups.'<br>';
-$dropgroupsresult = mysql_query($dropgroups, $linkID) or die("Drop groups Failed: ".mysql_error());
+$dropgroupsresult = $linkID->query ($dropgroups);
 $creategroups="CREATE TABLE IF NOT EXISTS `group_permissions` (
   `group_id` mediumint(9) NOT NULL auto_increment,
   `group_name` varchar(100) NOT NULL default '',
@@ -68,9 +95,12 @@ $creategroups="CREATE TABLE IF NOT EXISTS `group_permissions` (
   UNIQUE KEY `group_name` (`group_name`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 AUTO_INCREMENT=1 ";
 echo $creategroups.'<br>';
-$creagegroupsresult = mysql_query($creategroups, $linkID) or die("Drop users Failed: ".mysql_error());
 
-
+if($linkID->query($creategroups) ===TRUE){
+  printf("Users created successfully.\n");
+}else{
+   printf($linkID->error ,"\n\n");
+}
 /*
 $query="INSERT INTO `content` VALUES (1, 0, 1, 'Home', 'Keywords', '[Cyberia] Lan Parties home page2', '<p>Welcome to [cyberia], the home of soon the be the best home grown lan community in South Australia. Our goal here is to provide a comfortable friendly environment where gamers and developers alike can get together and pit their skills against each othe', NULL, 20060213221026, NULL, 20060214184358)";
 $resultID = mysql_query($query, $linkID) or die("add content failed: ".mysql_error());
@@ -90,16 +120,16 @@ $resultID = mysql_query($query, $linkID) or die("add content failed: ".mysql_err
 
 $query="INSERT INTO `pages` VALUES (1, 0, 1, 'Home2', 'Site Home Page', 'home', 'admin', '2006-02-16 22:44:23', 'khendar', '2006-02-17 02:59:04', 'public')";
 echo $query.'<br>';
-$resultID = mysql_query($query, $linkID) or die("add content failed: ".mysql_error());
+$resultID = $linkID->query ($query);
 $query="INSERT INTO `pages` VALUES (2, 0, 2, 'About', 'Site About Page', 'about', 'admin', '2006-02-16 22:44:49', 'admin', '2006-02-17 02:54:20', 'public')";
 echo $query.'<br>';
-$resultID = mysql_query($query, $linkID) or die("add content failed: ".mysql_error());
+$resultID = $linkID->query ($query);
 $query="INSERT INTO `pages` VALUES (3, 0, 3, 'Links', 'Site Links Page', 'links', 'admin', '2006-02-16 22:45:08', 'admin', '2006-02-16 22:36:50', 'public')";
 echo $query.'<br>';
-$resultID = mysql_query($query, $linkID) or die("add content failed: ".mysql_error());
+$resultID = $linkID->query ($query);
 $query="INSERT INTO `pages` VALUES (4, 0, 4, 'Home', 'default description', 'default keywords', 'admin', '2006-02-16 22:45:18', 'admin', '2006-02-16 22:42:38', 'public')";
 echo $query.'<br>';
-$resultID = mysql_query($query, $linkID) or die("add content failed: ".mysql_error());
+$resultID = $linkID->query ($query);
 
 
 ?>
